requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
colorama>=0.4.4
lxml>=4.6.3
jinja2>=3.0.0
tldextract>=3.1.0
tqdm>=4.60.0
pandas>=1.3.0
urllib3>=1.26.0
fake-useragent>=1.4.0
openpyxl>=3.0.7
PySide6>=6.4.0
pyinstaller>=5.6.0
python-docx>=1.1.2
pywin32>=305
